Move HGCalRawDataDefinitions.h.
